change: export Base class from tests.common.testcase to .base
Export Base class from tests.common.testcase to tests.common.base to
simplify both of them and make easier for tests later.
